+++ scripts/ide-setup-identity.sh
@@ -26,8 +26,14 @@ cp -r "${ide_identity}/.ssh" "${ide_home}"
 # StrictHostKeyChecking no
 echo "StrictHostKeyChecking no
 UserKnownHostsFile /dev/null
+
 ForwardAgent yes
 Host git.ai-traders.com
+User git
+Port 2222
+IdentityFile ${ide_home}/.ssh/id_rsa
+
+Host gitlab.ai-traders.com
 User git
 Port 2222
 IdentityFile ${ide_home}/.ssh/id_rsa