+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * #8906 added test-kitchen tests and end_user tests
 * #8906 can be ci released
 * #8906 added gocd pipelines
+* #8832 `~/.ssh/config` contains git.ai-traders.com and gitlab.ai-traders.com
 
 # 0.0.1